CDH-46214. Do not ship HDFS/YARN client libraries with Spark.
Instead, pick them up from CDH, so that Spark 2 can benefit from
bug fixes and features when deployed on new versions of CDH.

Because parquet / avro are still bundled with Spark to avoid
binary compatibility issues, some changes to their scope in the
pom files were made, so that all needed jar files are packaged
in the Spark parcel.

This change re-introduces the classpath.txt file used in CDH 5.x
to track Hadoop dependencies; this is needed to avoid adding too
many files to the classpath and putting pressure on the default
fdlimit, since the classpath generated by CDH's configuration has
a lot of duplicate entries.