[CDH-46529] Fix backport of SPARK-15891.
This change fixes the way executor memory is passed to yarn -- we need
to convert "spark.executor.memory" to MB, since an "m" is automatically
appended to it to build the jvm args.  This fixes a bug in the backport
of SPARK-15891, because the original chance used a conf feature not
present in 1.6.