CLOUDERA-BUILD. CDH-44309. Add unit tests to build.sh
This commit adds a new parameter to build.sh that allows
for running unit tests via the same script. If the build
is being published, it would also tag with the build
with a tag (unit_tests_passed), implying that the unit tests
have passed. Note that the build is published regardless of
whether the unit tests pass or not. The default behavior is
to not run unit tests unless explicitly requested.