[SPARK-4105] retry the fetch or stage if shuffle block is corrupt
There is an outstanding issue that existed for a long time: Sometimes the shuffle blocks are corrupt and can't be decompressed. We recently hit this in three different workloads, sometimes we can reproduce it by every try, sometimes can't. I also found that when the corruption happened, the beginning and end of the blocks are correct, the corruption happen in the middle. There was one case that the string of block id is corrupt by one character. It seems that it's very likely the corruption is introduced by some weird machine/hardware, also the checksum (16 bits) in TCP is not strong enough to identify all the corruption.

Unfortunately, Spark does not have checksum for shuffle blocks or broadcast, the job will fail if any corruption happen in the shuffle block from disk, or broadcast blocks during network. This PR try to detect the corruption after fetching shuffle blocks by decompressing them, because most of the compression already have checksum in them. It will retry the block, or failed with FetchFailure, so the previous stage could be retried on different (still random) machines.

Checksum for broadcast will be added by another PR.

Added unit tests

Author: Davies Liu <davies@databricks.com>

Closes #15923 from davies/detect_corrupt.

(cherry picked from commit cf33a86285629abe72c1acf235b8bfa6057220a8)

CDH Note: To backport the patch to the CDH version that is based on the 1.6 line, some changes
from SPARK-13921 were included (notedly some helper classes and methods), and some minor changes
made to account for the shuffle encryption code in the CDH branch.

Cloudera ID: CDH-48790

(cherry picked from commit 0a542ed55331486b20724ef09dd0fedeba90b7d3)