CLOUDERA-BUILD. CDH-48784. Respect app's SparkConf in YARN cluster mode.
The YARN allocator in cluster mode was ignoring executor core and memory
configs set by the user code in SparkConf, choosing instead to respect
only the launch-side configuration. If there was no launch-side config,
the defaults were used instead.

This change fixes things to make YARN cluster also respect the app's
SparkConf, and gives preference to it over values set in the launcher
side. This matches the behavior of client mode.

This was fixed in 2.0 as part of SPARK-12343, but the full diff is
too risky for backporting to 1.6, so this change is a customized
version that only applies to executor cores and memory.

(cherry picked from commit c5b7149ab01adf0e722c2106d36d0ee5580380b3)
(cherry picked from commit 1ba181a8f39710382cf0987e4fdea278496a8f10)