CDH-49085. Encrypt cached data stored on disk.
Spark 1.6 has a slightly different block manager than 2.0, and cached
data is not automatically encrypted here when evicted to disk. This
applied to both deserialized and also serialized data (in 2.0, serialized
data in memory is also encrypted, making eviction easy).

This change makes the changes so that paths that evict blocks to disk do
proper encryption, and the read paths do decryption only when necessary.

The disk-related storage level tests in BlockManager suite were updated
to run with encryption both on and off, with some extra checks added for
the encrypted case.