CDH-46402. Create a custom hive-exec jar with reduced dependencies.
This change creates a Spark-specific packaging of the hive-exec artifact,
based on the "-core" version that does not shade dependencies. Spark still
needs to relocate Kryo classes for this artifact since Hive depends on a
different version than Spark's, but otherwise, the change relies on maven
to harmonize all the dependencies, and we end up with a slimmer hive-exec
that shares a lot of dependencies with Spark (such as Jackson and protobufs).

The only unusual thing here is that some transitive dependencies have to
be explicitly listed, otherwise they disappear from the dependency list
in the pom created by the maven shade plugin.

One thing maybe worth noting is that the new hive-exec jar does not include
any HBase classes. This is better than before, since we require users to
explicitly add HBase jars to Spark when they use HBase, and having them in
hive-exec was a potential source of issues.

(cherry picked from commit c7a927633b40210ac76a6f90527bff984092f8b2)