CDH-49608. Encrypt blocks stored in memory.
The underlying issue is that when fetching a remote block, the code does
not know where the block was stored in the remote block manager. So, with
the current 1.6 code, if the block was in memory, the reader would get
unencrypted data, but if it was on disk, the reader would see encrypted
data.

That means that the reader cannot know how to read that data. So, instead,
do what Spark 2.0 does and always encrypt bytes stored in the memory store,
so that fetching them will always return encrypted data, no matter which
code path is used. This adds some overhead, but avoids having to change the
RPC protocols to include information about whether the block being read is
encrypted or not, or changing the local read path of the block manager, which
would be even more expensive (i.e. would make it complicated to use memory
mapped files).

To properly support that a few places had to be changed. Namely:
- adding a block to the block manager encrypts the data
- when the bm replicates the block, it sends over encrypted data, so the remote
  manager doesn't need to encrypt it again
- evicting a block from the memory store to the disk store does not decrypt
  the data, the data is just dumped into a file
- the broadcast code needed slight corrections to handle encryption

The new "ByteBufferOutputStream" class was just backported from Spark 2.0 (it
was added for a different change unrelated to this feature, but is useful
here).

One thing that deviates from Spark 2.0 is that WAL files written by streaming
apps are not encrypted, and a small change had to be made to the WAL code to
account for that.