CLOUDERA-BUILD. CDH-46754. Call QueryExecutionListener callback methods for DataFrameWriter methods
Calls the QueryExecutionListener callback methods onSuccess() and onFailure() methods for DataFrameWriter output methods.  It fixes an issue where the callback methods were being called for several of the DataFrame methods like take, head, first, collect etc. but didn't get called for any of the DataFrameWriter methods like saveAsTable, save etc. Added a new property "spark.sql.queryExecutionListeners" that can be used to specify instances of QueryExecutionListeners that should be attached to the SparkSession when the spark application starts up.