CLOUDERA-BUILD. CDH-46754. Adding output table to LogicalRelation
Changes to support the output table in case data is being persisted as done in the PR for SPARK-12724. We don't need the exact changes in the PR but only the change related to the output table which is the 3rd parameter in the constructor definition of the class LogicalRelation. This is necessary to have the output data in the QueryExecution object for analysis. So you will see changes like "case LogicalRelation(r, )" changed to "case LogicalRelation(r, , \_)" with the 3rd parameter being the output table name.