CLOUDERA-BUILD. CDH-46754. Generate lineage data for Spark SQL applications
Addition of the lineage project which will house the lineage code responsible for parsing and outputting the lineage data to the navigator json file.
Lineage of Spark applications using Spark SQL queries will be outputted to the lineage log directory. For each spark sql query that has inputs or outputs for eg. from hive, parquet file, json file etc., the tables or files from where the data was read and to where it was written, the columns read, data source type of the inputs/outputs will be written out as lineage data.