CLOUDERA-BUILD. CDH-49619. OutputMetadata in spark lineage json should contain fully qualified database name
The fully qualified table name is not available in the QueryExecution object for output metadata. Navigator needs a fully qualified table name. This commit aims to expose the current database via the sql catalog. It has the partial changes from the PR for SPARK-13571. This PR was never accepted into Spark upstream since the work was done by another larger piece of work associated with changing interfaces.