CDH-49004. Package both Kafka 0.9 and 0.10 backends in Spark.
This change makes changes to the pom files so that Kafka 0.9
(based on the 0.8 code in Spark) and 0.10 backends are packaged
in separate directories in the Spark distribution. POMs were
adjusted so that duplicate jars (i.e. jars already part of the
main Spark distribution) are not re-packaged in the Kafka-specific
directories.

On top of that, a new CSD parameters allows the user to choose
the default version, which defaults to 0.9 (to maintain backwards
compatibility). A new environment variable can be used by users
to control which version gets added to the Spark classpath.

Kafka 0.8 support is deprecated as of Spark 2.3.0 and put behind
kafka-0-8 profile. This commit turns this profile on.

(cherry picked from commit 599be7fce7fd7707a3c5ef692cf45a1365070f87)