CDH-51334. Avoid explicitly setting the location of managed Hive tables.
When creating a table on a remote metastore, verify if the explicit table
locations set by Spark in some cases are the same as the ones the metastore
would create. In those cases, do not explicitly set the table's location,
to work around an issue we've found in Sentry.

(cherry picked from commit 5fb40d0344030efad24add8a7c8072dd1aea4e0b)