CDH-51636. Auto tag builds, add script to update versions.
Auto-tag builds based on the pom version.This tags non-snapshot
builds as "rc" and snapshot builds as "snapshot", which follows
our release workflow and will allow us better control of which
automated tests run for each kind of build.

Also add a script that automates updating the pom and CSD versions
to the "next" version. To cover all pom files using the maven version
plugin, I moved the "flume-assembly" module into a new profile,
disabled by default, instead of removing it from the build.

Finally, removed a redundant version in the CSD pom, since with the
maven version we use, the version is inherited from the parent.