[SPARK-12297][SQL] Hive compatibility for Parquet Timestamps
This change allows timestamps in parquet-based hive table to behave as a
"floating time", without a timezone, as timestamps are for other file
formats. If the storage timezone is the same as the session timezone,
this conversion is a no-op. When data is read from a hive table, the
table property is always respected. This allows spark to not change
behavior when reading old data, but read newly written data correctly
(whatever the source of the data is).

Spark inherited the original behavior from Hive, but Hive is also
updating behavior to use the same scheme in HIVE-12767 / HIVE-16231.

The default for Spark remains unchanged; created tables do not include
the new table property.

This will only apply to hive tables; nothing is added to parquet
metadata to indicate the timezone, so data that is read or written
directly from parquet files will never have any conversions applied.

Note that CDH-35305 is the backport of SPARK-12297 to cdh 5.12.
However, the code has diverged so much that its not even a cherry-pick +
conflict resolution, its a re-implementation.  The good news is, the
main logic is similar and not very complicated.  The main difference
from upstreeam is interaction w/ hive for metadata operations (eg.
create table).

Cloudera ID: CDH-35305