CDH-50462. Allow the RM proxy IP filter to be disabled.
Add a config option to disable the RM IP filter to work around an issue
with the weave overlay network, where all the source IPs are translated
to be weave's IP, breaking the filter.

The change also uses YarnProxyRedirectFilter to redirect the link from the
RM UI to the driver's address.